Combine trending graph into monolith
