+++ package.json
@@ -6,8 +6,7 @@
     "dev": "NODE_ENV=dev nodemon index.js",
     "dev:router": "rover dev --supergraph-config supergraph-config-dev.yaml --router-config router-config-dev.yaml",
     "start": "node index.js",
-    "subscription": "nodemon ./subgraphs/trending/subgraph.js",
-    "test": "npm run test:compile",
+    "test": "npm run test:compile && npm run test:compose",
     "test:compile": "tsc",
     "test:compose": "APOLLO_ELV2_LICENSE=accept rover supergraph compose --config supergraph-config-dev.yaml --output supergraph.graphql"
   },
