+++ subgraphs/trending/schema.graphql
@@ -1,5 +1,3 @@
-# noinspection GraphQLTypeRedefinition
-
 extend schema
   @link(url: "https://specs.apollo.dev/federation/v2.4", import: ["@key"])
